Texture viewer mode (#2)
Shows the textures for the current level in the main view. Cycles through the textures on key up. Renders the texture to a quad.

Prompt the user for a file with openfilename. Load that file in the library.

Interface for level and moved types somewhere that external things can get at them. Also made a concrete level type that implements the interfaces and makes a start on loading the level data.

#1