Disable 'flip' checkbox if there are no flipmaps in the level (#212)
Make it so checkbox can be disabled. This requires making it so that labels can change text colour.
Move colour to common library so graphics can use it.
Update references to ui::Colour to be just Colour.
Update includes to be <trview.common/Colour.h>
Check all rooms for alternate mode other than 'None'. If there are any, enable the checkbox, otherwise disable it.
Add check for _handles_input in the click calculations.
Store label in checkbox so text colour can be set later.
Disable the flip hotkey if there are no flipmaps.
Issue: #140