Create Resizer class (#218)
Create a WindowResizer class. This takes all the logic that was in trview.cpp and puts it in a class.
Also creates a trview.app project and trview.app.tests project. This means we can start adding tests for things in the main project (as they'll be moved from the main project into this one).
#217