Move 'recent files' to its own class (#221)
Move the 'recent files' functionality out of trview.cpp and into its own class in the trview.app library. The viewer now creates one of these and uses it directly.
Also add some tests in the trview.app.tests project.
Add trview.tests.common project. It currently has a function to create a hidden window to send messages to when doing testing. Deleted the version of this that was in trview.input.tests and updated tests to pass in a window name.
Issues: #219