Make a shaders project to build the shaders (#227)
Shaders are built in this project - if you do a full build/rebuild, the trview project will copy the output of the trview.shaders project (the cso files) into the resources directory of trview. Then, when trview builds the rc file references these files and builds them in to the exe. Default shaders extract them from there.
I don't like the pre-build step, but it's ok for now as it lets it run as a single exe file with no supporting files.
#97