Move swap chain creation to a new class
In order to render to multiple windows, the swap chain creation must be separated from the creation of the device and context.
Create a class (DeviceWindow) that will have the render target and swap chain in it.
For now, the viewer will only have one of these, for the main window.