Split Device class up to allow for rendering to multiple windows (#252)
The items window will be a separate window, so we need to be able to render to it as well using the same D3D device. This takes out the swap chain and render target from the Device class and moves them into the DeviceWindow class. We can make one of these for each window we render to.

I also changed the WindowResizer class to detect whether the window size has changed and only raise the event if it has done. This means that some code can be removed from the Viewer class. As a result of this change I made a bunch of functions take a Size instead of width and height.

Can start working on the items window after this.

Issue: #251 - Item search/list window