Add tokens for event callback lifetimes (#258)
For the std::function version of Event +=, return a Token. When the token is destroyed, it will inform the event. The event will then remove the function from the listeners. This stops functions that have captured things that have been deleted from being called. This is more important now that the UI might actually create/destroy some things.

Anything that uses the function += needs to store the token, so created TokenStore which is basically just somewhere to put tokens. When it dies, it takes the tokens with it.

Also removed and ignored the compiled shaders.

This is preparation for #251