Display some items in the items list (#259)
Show a basic list of items. Currently shows item number, room and type. The type is not yet translated into a string (small medipack etc) but is the 'type id' of the item.
Adds the Listbox class that shows items in a grid with headers. The headers can be clicked on to sort the items according to that header and the grid can be scrolled with the mouse wheel.
Clicking on an item doesn't do anything yet.
Have also added an on_hierarchy_changed event to controls so that the UI renderer can notice when the contents of a control have changed.
Main window now doesn't zoom scroll when the mouse isn't over the window.
Add scroll to control so the list box can listen to it.
Issue: #251