Allow user to select items from the items window (#270)
The user can now click on the items in the items window and the viewer will focus on the item in question.
The items window is created by default, but you can make more of them via the Windows menu.
There is a 'track room' option on an items window - if this is selected, it will only show items in the current room.
Also adds a type name json file. This is missing a lot of entries for TR4 and TR5 and some work is required to improve the other game type names. This can be done separately though.
Issue: #251