Display the triggers that trigger an item in the items window (#304)
Show a list of triggers that trigger an item in the details panel of the items window.
This just shows the room number and the type of trigger. The trigger click event exists and is raised, but nothing is listening to it yet.

Trigger is a new class to be used in the application, a less raw version of TriggerInfo.
Moved a few types around to help with this. Moved SectorFlag into a namespace as it was clashing with the word Trigger.
Triggers are parsed by Level on load and stored - when items are generated they take note of which triggers reference them which goes through to the item window.

Part of issue: #284