Focus the viewer on the trigger position (#307)
When a trigger is clicked in the item details panel, focus the viewer on the location of the trigger.
Y position is calculated by picking - if there is nothing for the ray to hit, it will use the centre height of the room.
If 'Track Room' is on in the Items Window, it will be disabled. The currently selected item will be scrolled into view on the main item list. If you want to get back to the room with the item in you will need to click on the item in the items list again. This is OK for now but may need reworking, depending on how we feel about it after trying it out for a while.
Issue: #284