Click on an item in the viewer to select it (#325)
Adds the ability to click on items in the 3D view to select them. This requires picking against the meshes and bounding boxes for the entities.
Code is moved from room into mesh so this can be shared. Mesh bounding boxes are generated when the mesh is loaded and transformed by the entity to quickly eliminate a pick, saving picking against the triangles of the mesh.
Now that the user can select an item outside of the item window I have added the 'Sync Item' checkbox in each item window. If this is checked, two way communication with the 3d view is enabled:

If a user clicks an item in the 3d view, every item window with 'Sync Item' checked will have that item selected and details loaded.
If the user clicks an item in an item window and 'Sync Item' is checked, the 3D view will be updated to focus on that item. All other item windows with 'Sync Item' will also focus on that item.
Sync item is on by default, but if you turn it off you can have a window just to look at item details without actually changing the focus. This seems like a good model so I'll use it for the trigger windows.

Sprites currently pick against a bounding sphere, so if they are stacked on top of each other it may be difficult. I will make an issue to improve this (potentially by sampling the texture to look for transparency). In any case, you can select the item via the item list (track room will make it easier).

Issue: #309