Add clicking on triggers in the 3D view (#333)
Now renders clickable triggers in the 3D view. These appear as magenta volumes above the trigger position. They can be clicked as normal. This then selects the trigger in the trigger window.
Triggers have a lower priority than items for picking.
Triggers have a green highlight (to contrast with magenta and yellow for items).
A few mesh and texture related files have moved over to trview.app because Trigger lives there and needed to reference them. Things should be moving over to that library anyway over time.
Converted Trigger usage to be a unique_ptr in the owning location (level) and a pointer elsewhere.
Created a VertexShaderStore and PixelShaderStore to restore shaders after rendering the outline step - when there were two steps, it was breaking it by not restoring the shaders.
Trigger volume positions are generated by parsing triangulation data or the slope data, whichever is present.
Fixed an issue where sector was storing a reference to something that was later destroyed - this was previously not used.
Added a checkbox and a toggle mode for triggers - pressing T will toggle visibility.
Issue: #332