Use room bottom for corners if trigger is in on wall sector
If a trigger is on a wall sector, use the room bottom instead of the sector floor as the sector floor is -127.
Also move the filling of the corners array above the check for no floordata - this is not required now, but when sector rendering is added it will need to happen.
#350