Basic implementation of measuring tool (#355)
First version of the measuring tool. Adds measuring mode which can be accessed by pressing 'M' or by clicking the button in the tool area at the bottom of the 3D view.
Once in measure mode, the first click will start measuring. The second click will be the end point of the measurement. At the moment the line is represented by white cubes.
The pick tooltip shows the current measure distance until you end measurement, at which point a label appears at the centre of the measurement to show the distance.
The toolbar class just has buttons in it. It may change later.
Also in this PR ICamera is moved to the trview.app project so that it can be used by the measure tool. This meant a bunch of includes had to be updated.
Cube graphics and measure button will be improved in a future PR.
Issue: #336