Calculate room bounding box from room info
Use the room info instead of the room mesh data to generate the bounding box as some rooms don't have any geometry. This was the case for the room under the exit in Jungle.
Also change mesh bounding box to have 0 size instead of infinite if there are no vertices. This doesn't affect the room problem but would probably be a problem in the future.
#365