Add a compass/axis display to the viewer (#367)
Adds an axis display to the viewer. This shows the direction of the X, Y and Z axis relative to the current camera.
The axis display itself is rendered to a render target with its own camera that matches the main viewer camera. The user can click on the little cubes at the end of each axis to align the viewer camera to that axis.
Camera class was moved to trview.app as the axis display needed it.
Creating a cube mesh was moved to a function as it was also used in the measure tool. The windings were reversed as they were wrong after the change to stop inverting y.
Issue: #92