Improve the triggers filter
Change triggers filter into a list box.
Add 'clicked_off' - this is called on the focus control when another control is becoming the focus control.
Added rendering sort order option for child_elements - the ui and the renderer need things in opposite orders.
In control, stop as soon as a child control has handled something.
Issue: #371