Add the first version of the routing feature (#413)
Adds the first version of routing. The user can place waypoints by right clicking in the 3D view and selecting add waypoint. This can be on an item, trigger or just a point in space.
User can remove waypoints by right clicking on the waypoint in the 3D view and selecting delete.
The waypoint window shows the position, some information about what type of waypoint it is and allows the user to type notes.
The route can be exported to a file and imported from a file.
Graphics are terrible and further improvements are planned.
Issue: #163