Keep track of where neighbour generation has been
Remove the recursion from neighbour generation.
Keep track of processed rooms so we don't go back to them - this previously only checked against the previous room visited. This means there is less of a performance hit with high depth numbers, so increase the limit to 20. This can probably be increased later.
Issue: #32