Call begin on the device when beginning device window (#557)
This lets the device set the blend states. They were currently only being set when the main window rendered. The main window is only rendering when required now, so when a tool window was moved (resized), it was rendering without the blend states and make the white boxes.
Had to remove const on the graphics device through the call chain.
Bug: #556