Move keyboard input into a single place
It now checks if the focus control is visible and accepts input - if it does, then it will not pass the input to the camera and the shortcuts.
Remove the 'go_to_room' active and use 'is_input_active' as that is what it was really asking in the first place.