Correctly create bounding box when there are no vertices
There are some meshes with vertices but no rectangles or triangles.
At the moment these come through as having zero vertices.
In this case set the minimum of the bounding box to be zero - this was being done for the extent, but the minimum was still being set to -flt_max.
Bug: #575