Detatch LabelNode when Label is destroyed
LabelNode now listens to when labels are destroyed and will remove the reference to the label and no longer attempt to access it.
This stops a crash to do with the alternate group labels being destroyed.
Also altered TextArea to only add and remove lines if the line count is different - this makes the debug performance way better just because of the tooltip not regenerating all lines all the time.
Bug: #630