Stop adding X and Y coordinate value to UVs
For some reason I was adding the XCoordinate and YCoordinate values on to the UV coordinates for textures. This doesn't normally cause an issue because they tend to be 0, but in the winter level, they were 127. The renderer is set up to clamp UV coordinates, so they were just using the bottom right of the texture, which was a transparent pixel.

Also take fix from @sapper-trle for loading v130 sound maps.

Closes #620