Create an use ITriggersWindow interface
Not quite perfect as I am not sure what to do with the collapsible panel class - the inheritance graph is not as simple as I would like.
For now I have replicated it and just called through to it - maybe it will become more obvious when I move some of the other windows to similar interfaces.
TriggersWindowManager needs to be converted to be be able to create different types at runtime.
#740