Reset windows and route before loading the Viewer (#750)
The Viewer was being loaded before the various windows had received the updated items/triggers/etc lists. This meant that when the Viewer fired some events they were causing the windows to reference data that had been deleted.
The real problem is probably just storing raw pointers instead of a smart pointer or some kind of non-heap data structure, but that will need bigger changes. For now just move the Viewer load after resetting the windows and route.
Added a test to enforce this. It's a bit prescriptive, but it's better than nothing until I can go back and make everything a bit better.
Closes #749