Clear the selected trigger in the windows when the level changes
When the level changes go through all windows that store the selected trigger and clear the selection. This stops them referring to triggers that no longer exist.
Add tests for this in `TriggersWindowManager` and `RoomsWindowManager`. Only tests for this specific fix are added for `RoomsWindowManager` - they will be added separately as they would be too big for this change.
Closes #751