Add geometry and static meshes to on click pick filters
In the change to a flag based pick filter in #801 geometry was missed off the list, so click on rooms stopped working. Added geometry back into the filter and also added static meshes as they are effectively room geometry.
Added a test for this so they don't fall out again.
Closes #804