Render static mesh collision boxes (#817)
Renders white wireframe collision boxes around static meshes in a pass after rendering all rooms - this can be expanded in the future if bounding boxes for other entities need to be added.
A 'Bounds' checkbox is added to ViewOptions to toggle visibility. This defaults to off and isn't persisted across instances of the application.
Converts ViewOptions to DI so that ViewerUI can be properly tested. Mock added.
Adds a shared bounding mesh for all StaticMesh instances in the DI registration.
In order to test the rendering of the bounds in Level I had to mock ID3D11DeviceContext which was quite large - it will be useful for future graphics mocking though.
Closes #763