Add an option to create a Shortcuts instance for a specific window
Most shortcuts can use the application window, but if it is a new window it has to has its own instance. This mainly affects input at the moment so it will create one when it makes a new input.
Make the command index static so that if more than one shortcuts instance happens to be made for the same window (this is the case for the input fields on the camera position) they don't clash with other shortcut commands.
Closes #824