Fix scrolling for text areas
Don't allow the user to scroll if they can see all of the text in the current view.
Add a margin to the scrollbar blob so you can actually see which part is the blob and which part is the background.
Make scrollbar invisible when scrolling is not possible - it will come back when enough text is added.
Closes #818