Add 'mid waypoints' feature (#821)
Add the ability to add a 'mid waypoint' which will create a waypoint in the centre of the square that you click on. This only works on floors at present. They can be created using the right click context menu.
ContextMenu now has an interface and is created using DI.
Save and load normals for waypoints - this means that you can now place waypoints on walls and have them aligned to the wall instead of being vertical. This only applies to normals with no y component - waypoints on slopes will still be placed vertically.
Closes #730