Add 'Randomizer Location' option for waypoints (#842)
Add an option to the context menu to create a Randomizer Location waypoint type. This is only visible if the user has toggled Randomizer Tools from the settings menu. When a Randomizer Location waypoint is selected in the route window the Randomizer properties are editable.
If a route has Randomizer elements then there is an extra filter available in the export dialog box that will create a JSON file in Randomizer format of all Randomizer Locations in the route.

Other changes:

Stack layout updated to shrink as well as expand.
Fixed a bug where the route window didn't show waypoints in the list when there was only one waypoint.
Update to latest nlohmann::json to get ordered json support.
IDialogs::save_file/open_file changed to take multiple filters
Deleted old load_file function
Closes #841