Merge left/right panels for collapsible panels into one json file (#864)
Create the UI for collapsible panels (Items, Triggers, Rooms, Route) from just one json file each instead of the two files before. This will allow for top level elements such as bubbles and tooltips to be created in the json instead of in code as they are now.

Changes:

Merge left/right panel json files for all windows
Add auto_size support - only for height at the moment. This will make the element expand down to the height of the parent element. This attribute is added to several places in the windows to make them auto resize
Remove lots of code from CollapsiblePanel that was creating and managing the panels.
Don't raise on_size_changed if the size set is the same as the current size - this stops an event loop in auto_size
Closes #863