Use toggle finding system for view options
Instead of having specific event handlers, getters and setters for checkbox view option settings have a system that finds checkboxes and raises the name when they are changed. Then eventually the `Viewer` can decide what to do based on the value received.
Still needs names to come from a constant as well as some way of hiding values based on the level and some way to raise number values.
#866