Fix definition of on_scalar_changed handler to take integers (#877)
The handler attached to `on_scalar_changed` was unfortunately defined with `bool` instead of `int32_t`, so the values were being converted to `bool` and so were always `0` or `1`. This broke the depth setting from the view options.

Added a test to stop this happening again which should have been there in the first place really.

Closes #876