Floordata window
Room sync for floordata window

Start trying to do sector highlight

Try and rework picking to not always be picking

There was a sort of bug where if the cursor was outside the window then it would pick with empty, but then stop was not true so it would clear the minimap selection. I think picking might be too complicated and need reworking slightly.

Start on selected sector

Highlight sector properly

Highlight sector when hovering in floordata
Also do it for both views.

Remove noise

Don't repeat sector and room information