Fix event handler unregistering ownership loop
In the various window managers the `shared_ptr` for the newly created window was being captured in a lambda so that it could be removed from the window collection on an event when it was closed. Unfortunately capturing the `shared_ptr` meant it outlived where it was stored, so there was a crash.
Change to capture a `weak_ptr` instead - it should have been this in the first place as the `_closing_windows` field is a `vector` of `weak_ptr`s anyway.
Closes #919