Fix hide checkbox breaking window list sorting
When an item/trigger/light was made invisible previously the window manager was passing the entire list again and this was replacing the current list. This meant that sorting was lost and also the list was being updated in the middle of an iteration, so there was a strange ordering in the output.

Change `update_items` to pass only the item that has changed and call it `update_item`. This updates one item and doesn't re-sort the elements. Remove this function from Triggers/Lights windows as they work with `shared_ptr` unlike `Item`.
Closes #952