Fix Go To window not raising event on index 0 (#969)
If the user entered 0 as the index in the GoTo window ImGui would not return true for that - so the on_selected event was never raised. Instead we also raise the event in the section that closes the dialog on escape/enter - it is still raised in the InputInt check as the +/- arrows trigger that path.

Also change ViewerUI to pass the currently selected room/item as an initial value when making the GoTo window visible. This is a bit nicer than it always being at 0.

Bug reported by @TempleOfHorus

Closes #968