Add 'reload' function (#978)
Add a function to reload the current level. This can be triggered from the file menu or by pressing the F5 hotkey.

As part of this I have merged LevelSwitcher, RecentFiles and the FileDropper into a single FileMenu class as managing all three was getting complicated. The new class isn't too big and it cuts down on the number of events that have to be subscribed to by the application as well as removing a few constructor parameters.

Closes #976