Add Alternate and Alternate Group filters to room window (#1030)
Add filters for Alternate and Alternate Group to the Room window.
Since these are optional attributes added the ability for a single getter filter to have the is present and is not present ops - as long as there's a predicate. There was also a bug where the predicate wasn't being used which I fixed.
The ability to click the Alternate field and go to that room was missing - might have been lost when moving to ImGui. I have added it back in.
Closes #1029