Clear the log before loading levels
Since the level loading is currently the only thing that writes to the log, clear it before loading a level. Without this we just amass more and more log messages which takes a lot of memory after a while. Will fix log system to not do to this in the future.
Closes #1043