Sort window lists on initialisation (#1049)
When a window gets new contents given to it by the outside force a sort of the contents. This means that windows will keep their filters and sorts through level switching. Sorts will also survive closing and reopening the program. Also preserve command trigger filter in the triggers window.
Also fix sorting on windows that sorted on a type enum - they now sort on the string instead of the enum value.
Closes #1048