Only apply transparency in transparent object rendering (#1064)
In some older customs the texture tiles had 00 as the alpha component. When doing the normal non-transparent rendering trview was still using this alpha, so now set alpha to 1.0 unless we're rendering the transparent elements.
Apply a fix for when geometry has a texture index of 0x7fff. In this case, use the last known good texture index.
Add bounds check for UVs.
#1050