Pass level to room on creation
Pass level to the room source on creation - this is so the room knows about its parent without the previous reference - with Lua this could expire if someone was holding on to a room. May need to move to contruct/init for room too as it has sectors to create.