Lua: Initial UI bindings and plugin system (#1138)
Add the initial ImGui bindings. These aren't complete but have been enough to make the monkey finder example plugin. More will be added in batches.

Add a way of unregistering from an event handler - this is required as trview hasn't had levels surviving and coming back before (Lua can now keep them alive).

Lua:

Add selected_item property to level.
Add local_levels property to trview.
Change trview:load to be a coroutine - this allows for plugins to load levels in the background. A mutex was added to log as otherwise this was causing crashes.
Plugins can now have render_toolbar and render_ui functions that the application will call if they're present. UI is for rendering regular windows and toolbar is for rendering a button into the bottom toolbar.

#1129