Rework Route I/O and Route Window for Randomizer (#1154)
This started off as adding the route I/O for the Lua bindings but grew to a larger rewrite of the Randomizer integration and the Route window in general. The import and export buttons are now gone, replaced with a menu bar. Routes now keep track of where they were saved and know what type they are so the user doesn't have to repeatedly specify the file to export to.

- Added `RandomizerRoute` type which is a collection of waypoints. It swaps the waypoints out when the user changes the opened level. This lets the user use one set of Randomizer locations without having to open it manually for each level.
- Added menu bar to Route Window
- Replaced `Import` with `Open`
- Replace `Export` with `Save` and `Save As`
- Added `Reload` command
- Added option to create a new `Route` or `RandomizerRoute` in the Route Window
- Randomizer settings are now only shown for `RandomizerRoute`
- Route Window now shows level selection for `RandomizerRoute`
- Default route type now depends on whether Randomizer tools are enabled
- Added Lua bindings for `open`, `save`, `save_as`

Closes #1149 
Closes #1155