Allow Sector to be used to create Waypoint (#1167)
Let user pass in a `Sector` when creating a `Waypoint` using `Waypoint.new`.
This will create a mid waypoint at the provided sector.
Also has a couple of performance adjustments; the randomizer route now just adds the waypoint to the list when a waypoint is added instead of regenerating the whole thing. `Waypoint` and `Sector` now no longer store instances in a map, they just create the shared_ptr in userdata directly. Other classes will need to do this too.
Closes #1159