Fix issue with spaces in paths on open with
When trview is associated with level files and you open a level in explorer if the path has spaces in then the `CommandLineToArgvW` will split on spaces - we now stick them back together to form the original path.
This means flags have to come first but they aren't really used yet anyway.
Also fixes an issue in the map renderer where it would try to use imgui before it was ready in some situations.
Closes #1192