Add Categories property to Items (#1202)
Add a `Categories` property to `Item`. The user can see this category in the items window and is able to use filters against the category property. This will help for finding all pickups, all enemies or for excluding a certain type of item amongst other uses.

The `Categories` field is accessible from Lua so scripts can alter the categories on an item. The categories are taken as a copy of the initial category values from the type list so they can be safely changed by a script.

Closes #1200