Clear minimap on level load (#1207)
The minimap wasn't being cleared on level load. VICT.TR2 made this obvious as it had no rooms to to be loaded, so the old room display was still there.

Now when the level loads the minimap is cleared. Some functions to access room info directly from the level have been removed as they are no longer used. It will now take it from the `Sector` instead of using inidces.

Closes #1206