Add record count to filterable lists in windows (#1220)
Add a counter to show how many records are currently visible and how many there are for the main list in the various windows. If all records are visible it will show `y records`. If there is a filter it will show `x of y records`. `records` is replaced by the appropriate name in each window.

![image](https://github.com/chreden/trview/assets/4263940/70f9d868-7a43-4964-bef5-f69fa096ed2d)

Closes #1219