Allow user to change fonts (#1236)
Allows the user to change the fonts used in the program. Current options are for:
- Default - pretty much everything
- Console - the console window
More can be added later.

Fonts are now retrieved by key and not stored by consumers so they can be switched without consumer having to deal with that.

Windows are adjusted to better handle resizing. The main tables will adjust for font sizing. The `X` button to delete a save in a route waypoint has been removed and replaced with a right click context menu.

Switches to using FreeType rendering for fonts as with the STB implementation loading some fonts would crash the application - the FreeType implementation lets us catch the errors a little better.

Closes #970