Clear scroll_to variables after room content tab rendered
Previously the rendering of the selected entry would clear the `scroll_to` flag - however if there were no items (for example) in the room there would be nothing to clear the flag so it would repeatedly select the tab. Resetting the flag after the tab is rendered as well covers this.
Closes #1244