Add background loading (#1257)
Load the next level in the background. Show progress at the top of the screen and switch to the next level once loading is complete. This stops the program from hanging during level loading which is particularly noticeable with custom levels.

Texture loading is also changed so that levels now output the textures as they are loaded (where possible - can't be done with TR4 at the moment due to the compressed data). This should cut down on the peak memory usage when loading a level as the raw data is discarded once the D3D texture has been created.

#1186