Change picking to stop using indices (#1364)
Change `PickResult` to just have a `weak_ptr` to the thing that was picked instead of an index used to retrieve it from the level.
As part of this it made sense to change visibility to be a direct call instead of flowing events up and down.
Also fixed an issue with the auto-hider where it was applying auto-hide even if it was disabled.
Closes #1269