Add filter for 'Extra' value to Triggers window
Add a multi-getter trigger to the triggers window that filters on the 'Extra' data that cameras and a couple of other commands have.
Since the triggers window is filtering on triggers and not on commands all of the extras are joined and that is what is queries - for example if a trigger has two commands, one with 255 and one with 256 the trigger will have extra value of [255, 256] and will match both.
Closes #1409