Improve version detection, transparency in OPSM1 (#1431)
Improve the version detection code for PlayStation levels specifically to check the actual version number instead of relying on the length of the sound header. This means seeking in to the right place in the file.

Also fix some issues with transparency and lighting on OSPM1 (Version 38).

Start on the versions document.

Closes #1429 
Related to #1427 and #1430