Add a reusable model class (#1436)
Move the 'model' (collection of meshes) logic out of the `Item` class and into the new `Model` class. Stored in `ModelStorage` the `Item`, they are all generated on level load and requested by `Item`.

This can be used next for the model viewer.

Also when doing this a lot of the references to `ITextureStorage/ILevelTextureStorage` have been removed as it can be provided directly to the mesh on creation via the mesh source.

Related to #1435