Enable animated textures in 3D view (#1453)
Load and display animated textures in the 3D view.
Instead of adding yet another triangle type this change consolidates `TransparentTriangle` and `CollisionTriangle` into `Triangle`, adds animation elements and skips the direct use of vertex and index buffers until inside the mesh.
As part of this remove the lazy scene rendering as the scene will be changing all the time anyway - this means a lot of events get removed.
UV rotation textures are partially implemented but aren't currently used as it requires parsing the script to be able to determine the correct behaviour. Some TR4 and 5 levels will have water that isn't currently animated for this reason.
Closes #1454