Switch to point filtering to fix seams issue (#1460)
Switch from linear to point filtering for textures to finally fix the issue of texture seams.
As well as doing this it now generates a texture for each object texture - this is quite a lot of textures but they're all small, does mean more switching though but it stops any bleeding and will be useful for UV rotate.

This also removes the UV rotate texture splitting because it may be causing issues with subresource copying failing sometimes.

Closes #1459