Re-add linear filtering as an option (#1473)
With the recent change to split all textures into their own texture (not using them directly from the textile) it turns out that linear filtering works a lot better than it used to as long as it is on clamp.
Add a settings window option to enable linear filtering, disabled by default.
Remove the UV tweaking from before the replacement textures were used.
Closes #1472