Select sector in room window when selected in diff (#1489)
When user clicks on a sector in the diff window select that sector in the rooms window so the floordata mode can be more easily used to know exactly which sector it is that has changed.
Also update the windows SDK because the GitHub agents don't have the current one.
Closes #1340