Handle multiple portals on a sector
Sometimes sectors can have several portals on a sector - forward and up usually. Previous the second portal was overwriting the original. This changes it to store all of them and act on the 0th one - this is usually the forward portal. At some point will rewrite the minimap to be able to display the extra portals.

This won't fix geometry mode issues that were caused by the overwritten portals, that is in a separate issue.

Closes #1496