Convert UserSettings passing to messaging (#1507)
Convert setting passing to a message based system instead of the events.
Recipients can ask for settings via a message and something (`Application`) will answer with the message.
Message system is pretty basic and not very good at the moment, but it's functional.
Also had to rename the old log `Message` to `LogMessage`.
Closes #1506