Ignore unused texture vertices, deduplicate replacement textures (#1510)
Ignore the unused vertices in object textures so the minimum x and y coords are actually based on the minimum used values. This makes the cut out textures much smaller.

Also deduplicate these textures - this cuts the number of replacement textures in half more or less - down from ~1600 to ~700 on The Great Wall.

Closes #1509