Show level names in level info and recent files (#1452)
Make best effort to show the in-game name of the loaded level.

Approaches included:
- TOMBPC.DAT if present for 1-3
- ENGLISH.DAT if present for 4-5
- STRINGS.TXT for remasters if present
- TR1/2X gameflow/text files if present
- Last resort hash list for other level names

Includes an update to nlohmann json to be able to cope with the trailing commas in the TRX json files. All json parsing has been given extra flags to allow for comments and trailing commas.

Level names are shown in the recent files display and the level info. Sorting mode can be selected in the settings window.

The hashes list is incomplete but will be added to.

Closes #1394